requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
django-debug-toolbar==5.1.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
hiredis==3.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
Unidecode==1.3.8
uritemplate==4.1.1
